Splits GUI code into multiple files
